mtd: peb: Do not mark as bad in mtd_peb_torture()
Both the Kernel and mtd-utils have peb torture functions and both
do not mark the block as bad automatically. Instead, the caller
must mark the block as bad when -EIO is returned from the torture
function. Do the same in barebox. This is necessary as the UBI code
otherwise may mark a block as bad twice: Once indirectly in
mtd_peb_torture() and then directly afterwards.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>