mtd: nand-mxs: Make ecc strength configurable via device tree
According to the binding doc the mxs NAND driver supports the
"nand-ecc-strength" and "nand-ecc-step-size" options. This adds support
for these options to the driver. The "nand-ecc-step-size" is not
really configurable, the only accepted value is 512 so this is merely
to sanity check that there's nothing specified that we can't yet
support.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>