crypto: caam - Always do rng selftest
The caam rng selftest is known to be broken in several i.MX
incarnations. To be on the safe side just unconditionally execute
it rather than trying to guess from HAB failure events if this is
necessary.
We can only do the selftest once per boot though, doing it a second time
yields an error:

rng_self_test: Job Error:
2101000.jr0@1000.of: 20001953: CCB: desc idx 25: RNG: Instantiate

so only do the test when rng is not yet initialized as tested with the
RDSTA_IFx status bits.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>