riscv: use generic bitsperlong.h
Linux highly depends on the fact that the 'long' and the pointer
have the same width, and so does barebox.

So, we can always use include/asm-generic/bitsperlong.h, which
determines BITS_PER_LONG depending on CONFIG_64BIT.

This is what Linux does (at least in the kernel-space), and barebox
can follow it.

Currently, barebox only supports 32-bit riscv, but this should work
when it supports 64-bit by adding CONFIG_64BIT to arch/riscv/Kconfig.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>