mips: use generic bitsperlong.h
Linux highly depends on the fact that the 'long' and the pointer
have the same width, and so does barebox.

So, we can always use include/asm-generic/bitsperlong.h, which
determines BITS_PER_LONG depending on CONFIG_64BIT.

This is what Linux does (at least in the kernel-space), and barebox
can follow it.

It is true that MIPS Linux references _MIPS_SZLONG
(arch/mips/include/uaspi/asm/bitsperlong.h), but this is bacause
the user-space cannot reference CONFIG options. For the kernel-space,
it uses the generic definition from include/asm-generic/bitsperlong.h.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>