Documentation: zii-vf610-dev: Fix 'halt' in openocd configuration script
When attempting to use the existing configuration file to bootstrap
an SCU4 AIB board, openocd would consistently fail on the 'halt' command
with a "Target not examined yet" error.  Dumping target information
showed that vf610.cpu1 was the active target, rather that vf610.cpu0.

Adding a command prior to the halt to explicitly set the target core
resulted in proper operation.

Tested with openocd-0.10.0+dev-00924-g16496488 on a Zodiac SCU4 AIB.

Signed-off-by: Cory Tusar <cory.tusar@zii.aero>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>