smc911x: fix inverted poll-for-ready
The ! operator is applied wrongly to only smc91xx_reg_read's return value,
when probably the intention was for it to apply to the whole expression.

However, wait_on_timeout keeps looping while the condition is false,
so dropping the ! is the right thing to do. Do the right thing.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>