sandbox: compile with symbol -fvisibility=hidden
barebox defines many symbols that are reserved for the C implementation.
This collides with the libc, when building barebox for ARCH=sandbox.

Specify -fvsibility-hidden, so libraries barebox is linked against don't
inadvertently use barebox' functions.

This fixes a heap corruption occurring when issuing fb0.enable=1 on my
system, because X11 used strdup out of barebox, but free out of glibc[1]:

     binding file /lib/x86_64-linux-gnu/libX11.so.6 [0] to
     ./barebox [0]: normal symbol l `strdup' [GLIBC_2.2.5]
     [...]
     binding file /lib/x86_64-linux-gnu/libc.so.6 [0] to
     /lib/x86_64-linux-gnu/libc.so.6 [0]: normal symbol `free' [GLIBC_2.2.5]

[1]: cf. CFLAGS +=-Dfree=barebox_free in arch/sandbox/Makefile

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>