Documentation: zii-imx7d-dev: Fix errors with openocd configuration script
When attempting to use the existing configuration file to bootstrap an
RMU2 board, openocd would consistently fail in the 'safe_reset' function
with a number of errors.

Disabling the internal reset handling (similar to what's done on the
i.MX7 Sabre boards) allowed barebox to download and run successfully.

Tested with openocd-0.10.0+dev-00924-g16496488 on a Zodiac RMU2.

Signed-off-by: Cory Tusar <cory.tusar@zii.aero>
Tested-by: Andrey Smirnov <andrew.smirnov@gmail.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>