ARM: dts: imx25-karo-tx25: include Linux dts instead of duplicating it
This reduces duplicated information and benefits automatically from some
changes that were done in Linux since the board was last touched in
barebox. Also the intended differences are more obvious this way.

To not include dts/src/arm/imx25.dtsi twice (which undoes some intended
changes) it is necessary to drop the respective include from
arch/arm/dts/imx25.dtsi. This is not a problem as tx25 is its only
consumer.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>