add of_diff command
The of_diff command compares two device trees against each other
and prints a diff-like result. This can be handy to find out the
differences between the barebox live tree and the one we start the
kernel with. Another usecase would be to examine the changes our
of_fixup process introduces (of_diff - +)

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>