Documentation: make gen_commands helper python3 compatible
On some machines the python command is provided by Python 3 while on
most (at least in my bubble) it is still Python 2. Modify the code to
make it usable by both Python versions.

print_function is available in __future__ since Python 2.6.0a2, which
shouldn't be a relevant restriction.

The modified script generates the same documentation as the old one;
independent if the script is called using Python 2 (here: 2.7.16) or
Python 3 (here: 3.7.3).

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>