blspec: skip entries that are not blspec entries
The list of boot entries can contain items that are not blspec entries.
For example, the bootchooser also adds an item to this list. Therefore,
blspec cannot unconditionally interpret entries as blspec entries.

The error is reproduced by listing the boot entries with the bootchooser
and a rootfs with blspec entries, e.g.:

	boot -l bootchooser /mnt/nfs

Check if a bootentry is a blspec entry by testing if the boot function
is blspec_boot.

Reported-by: Thomas Hämmerle <Thomas.Haemmerle@wolfvision.net>
Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>