efi: retire efi_compare_guid in favor of efi_guidcmp
Both functions wrap the same memcmp except that one uses pointers to
GUIDs and the other passes the GUIDs by value.

The function is static inline, so it doesn't really matter which one we
keep. We'll drop efi_compare_guid because it's been used once only in the
code base so far.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>